+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the micro-web-service project will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+[comment]: <> (## [Unreleased])
+## [4.1.0] - 2021-04-11
 ### Added
 - 'patch' to allowed HTTP methods.
 
@@ -59,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial code.
 - Setup for this project.
 
-[Unreleased]: https://github.com/Ionaru/micro-web-service/compare/4.0.0...HEAD
+[Unreleased]: https://github.com/Ionaru/micro-web-service/compare/4.1.0...HEAD
+[4.1.0]: https://github.com/Ionaru/micro-web-service/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/Ionaru/micro-web-service/compare/3.0.2...4.0.0
 [3.0.2]: https://github.com/Ionaru/micro-web-service/compare/3.0.1...3.0.2
 [3.0.1]: https://github.com/Ionaru/micro-web-service/compare/3.0.0...3.0.1